+++ examples/spi_slave.rs
@@ -24,13 +24,15 @@ fn main() -> ! {
 
     let gpioa = p.GPIOA.split();
 
-    let sck = gpioa.pa5.internal_resistor(Pull::Up);
-    let miso = gpioa.pa6.internal_resistor(Pull::Down);
-    let mosi = gpioa.pa7.internal_resistor(Pull::Down);
+    let sck = gpioa.pa5.internal_resistor(Pull::Up).into();
+    let miso = gpioa.pa6.internal_resistor(Pull::Down).into();
+    let mosi = gpioa.pa7.internal_resistor(Pull::Down).into();
 
     // clock speed is determined by the master
     let nss = gpioa.pa4.internal_resistor(Pull::Up).into();
-    let mut spi = p.SPI1.spi_slave((sck, miso, mosi, Some(nss)), MODE);
+    let mut spi = p
+        .SPI1
+        .spi_slave((Some(sck), Some(miso), Some(mosi), Some(nss)), MODE);
     // alternativelly you could use software `chip select`
     // let mut spi = SpiSlave::new(p.SPI1, (sck, miso, mosi, None), MODE);
     // spi.set_internal_nss(false);